Add `trailing: 'always'`
The current behavior of `leading: true, trailing: true` is surprising to
some people. When `leading` is false, the semantics are that there will
always be *some* call to the debounced function within the debounce
interval of a call to it. But turning on `leading` changes that.

In case the current behavior is desirable to some users, this change
leaves current behavior alone and adds a new `trailing: 'always'` option
value.

Fixes #43.